Fixes #37084 - Drop simplecov dependency (#10837)
This is a test dependency, but marked as a development dependency. We
currently install all development dependencies while testing katello,
but don't do that for any other plugin. This is an effort to reduce the
development dependencies to pure development dependencies.

(cherry picked from commit 2ab734fd4cff41204742cefedeee0446895286d1)