[PR] Fixes #37279 - Re-create remotes when the type needs to change
When updating a remote, we may need to re-create it instead it the type
has changed from uln to rpm or vice versa.

For operations on a remote href, the type is now determined from the
href, rather than indirectly from the url. This is more robust since it
is the actually relevant source of truth.

Upstream PR: https://github.com/Katello/katello/pull/10942

Closes OR-4583