Fixes #38721 - Allow env_id param only on rolling CVs (#11489)
This fixes an issue introduced with the API validation that will reject
requests for non-rolling CVs if environment_ids are supplied. The issue
is that the UI makes such now failing requests! Since we are introducing
the API validation and this fix together we never had this issue in
orcharhino and do not need an internal ticket!

(cherry picked from commit d79a20324d1e4ec73b959916018751be9f677117)